requirements.txt
Django==4.2.3
django-debug-toolbar==4.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
